Step 8a: introduce controllers/ submodule and PolicyProtocol
Lays the groundwork for Step 8 without changing any behaviour. The
existing Controller class moves out of the top-level
holosoma_inference/controller.py into holosoma_inference/controllers/
controller.py; the top-level file becomes a deprecation shim that
re-exports the public symbols.

Adds two new files:

  controllers/protocol.py — PolicyProtocol (5 members: act, on_activate,
    on_deactivate, apply_velocity, apply_command) and the Command
    dataclass returned from act().
  controllers/__init__.py — public surface for the submodule.

No callers updated yet (they all still go through the deprecation
shim). 8b makes OnnxBasePolicy conform to the protocol; 8c rewrites
Controller to drive policies by protocol; 8d cleans up the legacy
flags.

Sim2sim harness: pelvis final=0.767 m, min=0.755 m (unchanged).
Pytest: 90 passed, 108 skipped, 0 failed.

Also updates docs/controller-design.md with the controllers/ layout
and adds PR_DESC.md listing the abstractions Step 8 will eliminate.
